requirements.txt
asttokens==2.4.1
autograd==1.6.2
autograd-gamma==0.5.0
backcall==0.2.0
blessed==1.20.0
blobfile==2.1.1
Brotli==1.0.9
certifi==2024.2.2
charset-normalizer==3.3.2
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
ecos==2.0.13
einops==0.8.0
entrypoints==0.4
exceptiongroup==1.2.1
executing==2.1.0
filelock==3.15.4
fonttools==4.53.0
formulaic==1.0.1
fsspec==2024.6.0
ftfy==6.2.0
future==1.0.0
gmpy2==2.1.2
gpustat==1.1.1
h5py==3.11.0
huggingface-hub==0.23.4
idna==3.7
importlib_resources==6.4.0
iniconfig==2.0.0
interface-meta==1.3.0
ipykernel==6.29.5
ipython==8.12.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter-client==7.3.4
jupyter_core==5.7.2
kaplanmeier==0.2.0
kiwisolver==1.4.5
lifelines==0.29.0
lxml==4.9.4
MarkupSafe==2.1.3
matplotlib==3.9.1
matplotlib-inline==0.1.7
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
mpmath==1.3.0
mypy==1.10.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
numexpr==2.10.0
numpy==1.26.4
nvidia-ml-py==12.560.30
opencv-python==4.10.0.84
openslide-python==1.3.1
osqp==0.6.5
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pip==24.0
platformdirs==4.3.6
pluggy==1.5.0
POT==0.9.3
prompt_toolkit==3.0.48
protobuf==5.26.1
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycryptodomex==3.20.0
Pygments==2.18.0
pyparsing==3.1.2
PySocks==1.7.1
pytest==8.2.2
python-dateutil==2.9.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
qdldl==0.1.7.post2
regex==2024.5.15
requests==2.32.2
safetensors==0.4.3
scikit-learn==1.3.2
scikit-survival==0.22.2
scipy==1.11.4
seaborn==0.13.2
sentence-transformers==3.0.1
sentencepiece==0.2.0
setuptools==69.5.1
six==1.16.0
stack-data==0.6.2
sympy==1.13.2
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tornado==6.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.42.3
triton==2.2.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
wheel==0.43.0
wrapt==1.16.0
yacs==0.1.8
zipp==3.19.2
